Draft 7

+++ app/src/main/java/com/example/infs3634app/Restaurants.java
@@ -6,8 +6,8 @@ public class Restaurants {
     private String name;
     private String cuisine;
     private String location;
-    private double myRating;
-    private double averageRating;
+    private float myRating;
+    private float averageRating;
     private int averageCost;
     private String address;
 
@@ -17,7 +17,7 @@ public class Restaurants {
 
     }
 
-    public Restaurants(String name, String cuisine, String location, double myRating, double averageRating, int averageCost, String address) {
+    public Restaurants(String name, String cuisine, String location, float myRating, float averageRating, int averageCost, String address) {
         this.name = name;
         this.cuisine = cuisine;
         this.location = location;
@@ -43,13 +43,13 @@ public class Restaurants {
 
     public void setAverageCost(int averageCost) { this.averageCost = averageCost; }
 
-    public double getMyRating() { return myRating;}
+    public float getMyRating() { return myRating;}
 
-    public void setMyRating(double myRating) { this.myRating = myRating; }
+    public void setMyRating(float myRating) { this.myRating = myRating; }
 
-    public double getAverageRating() { return averageRating;}
+    public float getAverageRating() { return averageRating;}
 
-    public void setAverageRating(double averageRating) { this.averageRating = averageRating; }
+    public void setAverageRating(float averageRating) { this.averageRating = averageRating; }
 
     public String getAddress() { return address;}
 
@@ -70,16 +70,16 @@ public class Restaurants {
     public static ArrayList<Restaurants> getRestaurants() {
 
         ArrayList<Restaurants> restaurants = new ArrayList<>();
-        restaurants.add(new Restaurants("El Jannahs", "Fast Food", "Kogarah",  4.50, 4.30, 20, "123 Smith Street"));
-        restaurants.add(new Restaurants("McDonalds", "Fast Food", "Auburn",  4.00, 4.20, 15, "23 Homer Street"));
-        restaurants.add(new Restaurants("EasyWay", "Drinks", "Randwick",  3.50, 4.00, 8, "11 First Avenue"));
-        restaurants.add(new Restaurants("Chatime", "Drinks", "Eastwood",  2.50, 3.50, 8, "22 Second Avenue"));
-        restaurants.add(new Restaurants("Tetsuya", "Japanese", "Sydney CBD",  4.50, 5.00, 300, "34 Third Avenue"));
-        restaurants.add(new Restaurants("Miyazaki", "Japanese", "Sydney CBD",  4.00, 4.50, 200, "55 Fourth Avenue"));
-        restaurants.add(new Restaurants("Golden Dragon", "Chinese", "Kogarah",  3.70, 3.20, 20, "66 Sixth Avenue"));
-        restaurants.add(new Restaurants("Fortune Cookie", "Chinese", "Hurstville",  4.60, 4.00, 15, "77 Seventh Avenue"));
-        restaurants.add(new Restaurants("Boganville", "Australian", "Punchbowl",  4.20, 3.50, 10, "88 Eight Avenue"));
-        restaurants.add(new Restaurants("YallahEats", "Lebanese", "Kensington",  4.90, 4.50, 10, "99 Ninth Avenue"));
+        restaurants.add(new Restaurants("El Jannahs", "Fast Food", "Kogarah",  4.50f, 4.30f, 20, "123 Smith Street"));
+        restaurants.add(new Restaurants("McDonalds", "Fast Food", "Auburn",  4.00f, 4.20f, 15, "23 Homer Street"));
+        restaurants.add(new Restaurants("EasyWay", "Drinks", "Randwick",  3.50f, 4.00f, 8, "11 First Avenue"));
+        restaurants.add(new Restaurants("Chatime", "Drinks", "Eastwood",  2.50f, 3.50f, 8, "22 Second Avenue"));
+        restaurants.add(new Restaurants("Tetsuya", "Japanese", "Sydney CBD",  4.50f, 5.00f, 300, "34 Third Avenue"));
+        restaurants.add(new Restaurants("Miyazaki", "Japanese", "Sydney CBD",  4.00f, 4.50f, 200, "55 Fourth Avenue"));
+        restaurants.add(new Restaurants("Golden Dragon", "Chinese", "Kogarah",  3.70f, 3.20f, 20, "66 Sixth Avenue"));
+        restaurants.add(new Restaurants("Fortune Cookie", "Chinese", "Hurstville",  4.60f, 4.00f, 15, "77 Seventh Avenue"));
+        restaurants.add(new Restaurants("Boganville", "Australian", "Punchbowl",  4.20f, 3.50f, 10, "88 Eight Avenue"));
+        restaurants.add(new Restaurants("YallahEats", "Lebanese", "Kensington",  4.90f, 4.50f, 10, "99 Ninth Avenue"));
 
         return restaurants;
     }
